Symbolic Structs

It is occasionally useful to represent structs symbolically, and retain the ability to perform getproperty on them. For example, let us consider the following struct representing a point in 2D cartesian coordinates.

using Symbolics, SymbolicUtils

struct Point2D{T}
    x::T
    y::T
end

To allow this struct to be used symbolically, the @symstruct macro must be used:

@symstruct Point2D{T}

Note that this macro requires that all type parameters of the struct be present in the macro use. With this, we can create symbolic structs like any other variable:

julia> @variables point::Point2D{Real}1-element Vector{SymStruct{Main.Point2D{Real}}}:
 point
julia> typeof(point)SymStruct{Main.Point2D{Real}}
julia> @assert point.x isa Num
julia> @assert point.y isa Num

Subsequently, fields of the struct can be accessed symbolically:

julia> point.x^2 + point.y^2point.x^2 + point.y^2

Array fields

What if we have a general point type capable of representing arbitrary dimensional cartesian coordinates?

struct PointN{T, N}
    x::Vector{T}

    function PointN{T, N}(x::Vector{T}) where {T, N}
        @assert length(x) == N
        new{T, N}(x)
    end
end

The standard @symstruct PointN{T, N} would work. However, Symbolics cannot infer the length of field x simply from the type PointN{T, N}. As a result, the field would be inferred as an unknown length vector, and allow arbitrary indexing such as point.x[1234]. To provide this information, the @symstruct macro allows specifying additional metadata:

@symstruct PointN{T, N} begin
    shape(:x) = (1:N,)
end

The shape option allows specifying the shape of array fields. Here, we're saying that the shape of field :x is (1:N,) (i.e. it is a 1-indexed vector of length N). The expression can use type information of the registered struct type. The specific details of all options are documented in the @symstruct macro. With the above declaration, we can now have safe symbolic PointN.

julia> @variables point::PointN{Real, 3}1-element Vector{SymStruct{Main.PointN{Real, 3}}}:
 point
julia> point.x[1]^2 + point.x[2]^2 + point.x[3]^2(point.x)[1]^2 + (point.x)[2]^2 + (point.x)[3]^2
julia> @assert size(point.x) == (3,)
julia> try
           point.x[4]
       catch e
           Base.showerror(e)
       endERROR: MethodError: no method matching showerror(::BoundsError)
The function `showerror` exists, but no method is defined for this combination of argument types.

Closest candidates are:
  showerror(::IO, ::UndefRefError)
   @ Base errorshow.jl:162
  showerror(::IO, ::Latexify.UnrepresentableException)
   @ Latexify ~/.julia/packages/Latexify/IJYMW/src/error.jl:24
  showerror(::IO, ::DimensionMismatch)
   @ Base errorshow.jl:170
  ...

Abstract types

Often, it is useful to specify that all subtypes of a given abstract type should be considered symbolic structs. The @symstruct macro can be used on abstract types to enable this behavior.

abstract type AbstractRecord{T} end

struct Record1{T} <: AbstractRecord{T}
    x::Vector{T}
    y::T
end

struct Record2{T} <: AbstractRecord{T}
    x::Vector{T}
    z::T
end

@symstruct AbstractRecord{T} begin
    shape(:x) = (1:3,)
end

As evidenced above, even options such as shape can be specified for abstract types. These are applicable to all subtypes.

julia> @variables rec1::Record1{Real} rec2::Record2{Real}2-element Vector{SymStruct}:
 rec1
 rec2
julia> @assert size(rec1.x) == (3,)
julia> @assert size(rec2.x) == (3,)
julia> rec1.x[1] + rec2.x[2] + rec1.y + rec2.zrec1.y + (rec2.x)[2] + (rec1.x)[1] + rec2.z

The @symstruct macro roughly follows Julia subtyping behavior. This means that more specific @symstruct declarations override less specific ones.

struct Record3{T} <: AbstractRecord{T}
    x::Vector{T}
    w::T
end

@symstruct Record3{T} begin
    shape(:x) = (1:4,)
end

struct Record4{T} <: AbstractRecord{T}
    x::Vector{T}
    f::Vector{T}
end

@symstruct Record4{T} begin
    shape(:f) = (1:5,)
end

julia> @variables rec3::Record3{Real} rec4::Record4{Real}2-element Vector{SymStruct}:
 rec3
 rec4
julia> @assert size(rec3.x) == (4,) # Different from the one declared by `AbstractRecord`
julia> @assert size(rec4.x) == (3,) # Falls back to the `AbstractRecord` definition
julia> @assert size(rec4.f) == (5,) # Uses the more specific declaration

Arrays of symbolic structs

Arrays of symbolic structs are required to have a concrete eltype

julia> @variables pts[1:3]::PointN{Real, 3}1-element Vector{Symbolics.Arr{SymStruct{Main.PointN{Real, 3}}, 1}}:
 pts[1:3]
julia> pts[1].x[1] + pts[2].x[2](pts[2].x)[2] + (pts[1].x)[1]

Nested structs

Symbolic struct types can be nested inside each other.

julia> @variables recs::Record1{Record1{Real}}1-element Vector{SymStruct{Main.Record1{Main.Record1{Real}}}}:
 recs
julia> recs.x[1].x[1] + recs.x[2].y.yERROR: FieldError: type Num has no field `y`, available fields: `val`

Registered functions of symbolic structs

There is some unique behavior when registering functions where arguments are (arrays of) symbolic structs. The @symstruct macro works by defining some interface functions to allow using the Symbolics.SymStruct wrapper type, similar to how Num is a wrapper type. All registration functions behave the same as they did prior to this feature if the SymStruct is unwrapped (using SymbolicUtils.unwrap). The difference in behavior arises from how wrapped SymStruct types are handled. In general, the @register_symbolic and @register_array_symbolic macros can only use information from the provided type annotations. They cannot know to declare methods for SymStruct{ConcreteFoo} if the function is registered with AbstractFoo. For example, the following syntax:

@register_symbolic foofn(foo::AbstractFoo)

Will only work for a variable declared as @variables foo::ConcreteFoo if @symstruct AbstractFoo is declared, regardless of whether @symstruct ConcreteFoo is declared or not. This will always work for @variables afoo::AbstractFoo.

@register_symbolic foofn2(foos::Vector{AbstractFoo})

Suppose we have declared @variables a::ConcreteFoo b::ConcreteFoo cs[1:3]::ConcreteFoo. foofn2([a, b]) will work only if @symstruct AbstractFoo is declared, regardless of whether @symstruct ConcreteFoo is declared or not. The same applies for foofn2(cs).

If you are not using symbolic structs, the registration macros behave exactly as they did prior to this feature.

API

Symbolics.@symstruct — Macro

@symstruct Foo{T1, T2, ...}
@symstruct Foo{T1, T2, ...} begin
  # options...
end

A macro which enables using type Foo with SymStruct as a symbolic struct. The first argument to the macro must be the struct type, with all type parameters named. The optional second argument is an optional begin..end block containing options that influence the behavior of the macro. The following options are allowed:

shape(:field) = # expression. For array fields, the shape of the field cannot be inferred from the type. In case the type of the field can be inferred from the type, it can be specified using this syntax. The expression must evaluate to an object of type Union{SymbolicUtils.Unknown, AbstractVector{UnitRange{Int}}, Tuple{Vararg{UnitRange{Int}}}}. The expression has access to the concrete type of the struct being accessed, with all type parameters available as declared in the first argument.

For example, given the following struct:

struct Record{T}
  x::Int
  y::Real
  z::T
end

It can be registered as

# Note: the type parameter must be declared, but the name itself does not matter
@symstruct Record{V} begin
# If `V` is an `AbstractVector` then field `z` is a 3-vector. Otherwise, it is a scalar.
  shape(:z) = V <: AbstractVector ? [1:3] : ()
end

Now,

@variables rec::Record{Int} rec2::Record{Vector{Int}}

rec.x, rec2.x will be Nums with symtype Int. rec.y and rec2.y will be Nums with symtype Real. rec.z will be a Num with symtype Int. rec2.z will be an Arr{Num, 1} with symtype Vector{Int} and shape [1:2].

In case the shape of a field is not provided, it will be inferred from the type. For AbstractArray subtypes, it will be SymbolicUtils.Unknown(ndims(arr_type)). Otherwise, it will be treated as a scalar.

Symbolics.SymStruct — Type

struct SymStruct{T}

Wrapper type for symbolic structs. Requires that the wrapped struct type T be registered with @symstruct. After registration, @variables can be used to create the symbolic struct.

# Here, `record` has type `SymStruct{Record}`
@variables record::Record

getproperty access on this is a symbolic operation, and returns an expression performing the appropriate field access. This can only wrap concrete struct types (isconcretetype(T) must be true). getproperty on this struct leverages fieldnames and fieldtypes. Thus, will thus not respect custom getproperty methods on the wrapped struct type.