Miscellaneous API

This page documents various utility functions and constants that don't fit into other categories.

Constants

Missing docstring.

Missing docstring for Symbolics.NAMESPACE_SEPARATOR. Check Documenter's build log for details.

Utility functions

Symbolics.linear_expansion — Function

(a, b, islinear) = linear_expansion(t, x)

When islinear, return a and b such that a * x + b == t. Instead of calling linear_expansion multiple times with the same x, prefer using Symbolics.LinearExpander.

Symbolics.LinearExpander — Type

struct LinearExpander

A functor which acts as the workhorse for Symbolics.linear_expansion. The operation linear_expansion(t::Symbolics.SymbolicT, x::Symbolics.SymbolicT) can equivalently be phrased as LinearExpander(x)(t). This functor caches data during processing, and thus several linear_expansion calls with the same x can be rephrased as multiple calls to LinearExpansion(x). This can help avoid repeated computation and allocations.